# docs/openapi/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIPP API",
    "version": "0.1.0",
    "description": "Generated by `npm run docs:generate` from JSDoc headers in app/api/ (see scripts/generate-openapi.mjs). Request/response schemas are not yet modelled — they will appear as zod-to-openapi definitions are added."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "BetterAuth session cookie. Set the cookie in Swagger UI using Authorize."
      }
    }
  },
  "paths": {
...
    "/api/admin/organizations/{id}/status": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "PATCH /api/admin/organizations/{id}/status",
        "description": "Super Admin only — change organization status with state machine validation.\nWrapped with wrapPiiRoute for payload encryption.\nState machine: PENDING → ACTIVE ↔ SUSPENDED → ARCHIVED (terminal)\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/[orgId]/status/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      }
    },
...
    "/api/admin/organizations": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "paginated organization list (page, pageSize, status, search)",
        "description": "GET — paginated organization list (page, pageSize, status, search)\n\nSuper Admin only. Writes are rate-limited per session. Cross-tenant reads\nand inserts run under a verified platform context (RLS Phase 3); creation\nis recorded in the audit log.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "create an organization (name, slug?, adminEmail?)",
        "description": "POST — create an organization (name, slug?, adminEmail?)\n\nSuper Admin only. Writes are rate-limited per session. Cross-tenant reads\nand inserts run under a verified platform context (RLS Phase 3); creation\nis recorded in the audit log.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      }
    },
    "/api/admin/organizations/search": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "GET /api/admin/organizations/search",
        "description": "Search organizations by name (prefix match, case-insensitive).\nRequires Super Admin authentication.\nWrapped with wrapPiiRoute for payload encryption.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/search/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      }
    },
...
    "/api/dashboard/admin/organizations/{id}/status": {
      "patch": {
        "tags": [
          "dashboard-admin"
        ],
        "summary": "PATCH /api/dashboard/admin/organizations/{id}/status",
        "description": "Update organization status (suspend/reactivate/archive).\n\nRequires super admin session.\n\nHandler: `app/api/dashboard/admin/organizations/[id]/status/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      }
    },
...
  }
}